+++ cloudpathlib/anypath.py
@@ -5,6 +5,7 @@ from typing import Any, Union
 
 from .cloudpath import InvalidPrefixError, CloudPath
 from .exceptions import AnyPathTypeError
+from .url_utils import path_from_fileurl
 
 
 class AnyPath(ABC):
@@ -22,6 +23,12 @@ class AnyPath(ABC):
             return CloudPath(*args, **kwargs)  # type: ignore
         except InvalidPrefixError as cloudpath_exception:
             try:
+                if isinstance(args[0], str) and args[0].lower().startswith("file:"):
+                    path = path_from_fileurl(args[0], **kwargs)
+                    for part in args[1:]:
+                        path /= part
+                    return path
+
                 return Path(*args, **kwargs)
             except TypeError as path_exception:
                 raise AnyPathTypeError(