+++ tests/test_anypath.py
@@ -1,3 +1,4 @@
+import os
 from pathlib import Path, PosixPath, WindowsPath
 
 import pytest
@@ -19,6 +20,29 @@ def test_anypath_path():
     assert issubclass(PosixPath, AnyPath)
     assert issubclass(WindowsPath, AnyPath)
     assert not issubclass(str, AnyPath)
+
+    # test `file:` scheme (only works with absolute paths; needs .absolute() on Windows)
+    assert AnyPath(path.absolute().resolve().as_uri()) == path.absolute().resolve()
+
+    # test file:// + multi arg
+    assert AnyPath(*path.absolute().resolve().as_uri().rsplit("/", 2)) == path.absolute().resolve()
+
+    # test no hostname
+    assert Path("/foo/bar") == AnyPath("file:/foo/bar")
+    assert Path("/foo/bar") == AnyPath("file:///foo/bar")
+
+    # windows tests
+    if os.name == "nt":
+        assert Path("c:\\hello\\test.txt") == AnyPath("file:/c:/hello/test.txt")
+        assert Path("c:\\hello\\test.txt") == AnyPath("file://c:/hello/test.txt")
+        assert Path("c:\\hello\\test.txt") == AnyPath("file:///c:/hello/test.txt")
+        assert Path("c:\\hello\\test.txt") == AnyPath("file://c%3A//hello/test.txt")
+        assert Path("c:\\hello\\test.txt") == AnyPath("file://localhost/c%3a/hello/test.txt")
+        assert Path("c:\\WINDOWS\\clock.avi") == AnyPath("file://localhost/c|/WINDOWS/clock.avi")
+        assert Path("c:\\WINDOWS\\clock.avi") == AnyPath("file:///c|/WINDOWS/clock.avi")
+        assert Path("c:\\hello\\test space.txt") == AnyPath(
+            "file://localhost/c%3a/hello/test%20space.txt"
+        )
 
 
 def test_anypath_cloudpath(rig):